% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcm.R
\name{as.PanelMLCM}
\alias{as.PanelMLCM}
\title{Structuring the panel dataset}
\usage{
as.PanelMLCM(y, x, timevar, id, int_date = NULL, y.lag = 0)
}
\arguments{
\item{y}{Numeric, the outcome variable.}

\item{x}{Matrix or data.frame of covariates to include in the model.}

\item{timevar}{The column containing the time variable. It can be numeric, integer or
a date object}

\item{id}{Numeric, the column containing the ID's.}

\item{int_date}{INSERT COLUMN TIME AND EXPLANATION (Optional)}

\item{y.lag}{Optional, number of lags of the dependent variable to include in the model.}
}
\value{
An object of class 'data.frame' and 'PanelMLCM'.
}
\description{
This function takes as input the panel dataset given by the user and changes the
ordering and the names of the columns to obtain an object of class 'PanelMLCM'
to be used by the function 'MLCM'.
}
\details{
This function is mainly for internal use of \code{MLCM}. It is exported to give
users full flexibility in the choice of the ML algorithms and during the panel cross validation.
See the documentation of the function \code{PanelCrossValidation}.
}
\examples{

# Start from a disorganized panel dataset
head(data)

# Run as.PanelMLCM
newdata <- as.PanelMLCM(y = data[, "Y"], timevar = data[, "year"], id = data[, "ID"],
                        x = data[, !(names(data) \%in\% c("Y", "ID", "year"))], y.lag = 2)

# Results
head(newdata)
}
