% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_cate}
\alias{boot_cate}
\title{Bootstrap inference for CATE}
\usage{
boot_cate(effect, cate, nboot, alpha)
}
\arguments{
\item{effect}{Numeric vector of estimated individual causal effects.}

\item{cate}{Object of class \code{rpart}, the estimated regression-tree-based CATE.}

\item{nboot}{Number of bootstrap replications.}

\item{alpha}{Confidence interval level to report for the ATE. Defaulting to 0.05 for a two sided
95\% confidence interval}
}
\value{
A matrix containing the following information: estimated CATE within
each node, estimated variance and confidence interval (upper and lower bound)
estimated by bootstrap. Each column corresponds to a terminal node of the tree.
}
\description{
Internal function, used within the MLCM routine for the estimation of CATE
standard errors and 95% confidence intervals in each terminal node of the tree.
It works by resampling the observations at each final node of the tree.
Note that the observations are the estimated individual
causal effects (computed by comparing the observed data with the ML predictions).
Our estimand of interest is the average of the individual effects, so at each bootstrap
iteration we average the individual effects, obtaining a bootstrap distribution for the ATE
(which is in fact a CATE as we do that in each terminal node, i.e., conditionally on covariates).
}
