% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcm.R
\name{MLCM}
\alias{MLCM}
\title{MLCM}
\usage{
MLCM(data, y, timevar, id, post_period, nboot = 1000)
}
\arguments{
\item{data}{A panel dataset in long form, having one column for the time variable, one column for the units'
unique IDs, one column for the outcome variable and one or more columns for the covariates.}

\item{y}{Character, name of the column containing the outcome variable.}

\item{timevar}{Character, name of the column containing the time variable.}

\item{id}{Character, name of the column containing the ID's.}

\item{post_period}{The post-intervention period where the causal effect should be computed. It must be contained in 'timevar'.}

\item{nboot}{Number of bootstrap replications, defaults to 1000.}
}
\description{
MLCM
}
