% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelcv.R
\name{PanelCrossValidationStag}
\alias{PanelCrossValidationStag}
\title{Panel Cross Validation Staggered}
\usage{
PanelCrossValidationStag(
  data,
  pcv_block = 1,
  metric = "RMSE",
  default_par = list(gbm = list(depth = c(1, 2, 3), n.trees = c(500, 1000), shrinkage =
    seq(0.01, 0.1, by = 0.02), n.minobsinnode = c(10, 20)), rf = list(ntree = 500), pls =
    list(ncomp = c(1:5)), lasso = list(fraction = seq(0.1, 0.9, by = 0.1))),
  trControl = NULL,
  ML_methods = NULL
)
}
\arguments{
\item{data}{A 'PanelMLCM' object from a previous call to \code{as.PanelMLCM}.}

\item{pcv_block}{Number of pre-intervention times to block for panel cross validation. Defaults to 1, see Details.}

\item{metric}{Character, the performance metric that should be used to select the optimal model.
Possible choices are either \code{"RMSE"} (the default) or \code{"Rsquared"}.}

\item{default_par}{List of parameters for the default ML algorithms. See the Details of the function
\code{PanelCrossValidation}}

\item{trControl}{Optional, used to customize the training step. It must be the output from a call to \code{trainControl} from the \code{caret} package.}

\item{ML_methods}{Optional list of ML methods to be used as alternatives to the default methods. Each method must be supplied
as a named list of two elements: a character defining the method name from all the ones available in \code{caret}
and the grid of parameter values to tune via the panel cross validation. See Details and the examples for additional explanations.}
}
\value{
For each intervention group, a list containing the best-performing ML method, the value of the given \code{metric} and the detailed results
        of all methods tried during the PCV routine (see also \code{PanelCrossValidation}).
}
\description{
Panel Cross Validation in a staggered adoption setting, i.e., when there are groups of units
treated at different times.
}
\details{
The function works in the same way as \code{PanelCrossValidation} as it repeats independently the PCV procedure for each group of unit defined by
their treatment date. Note that the \code{int_date} argument must be a column of \code{data}.
}
\examples{

### Example 1. Using the default ML methods in a staggered setting
set.seed(1)

# Assume the following intervention dates
int_year_i <- c(rep(2019, times = 60), rep(2020, times = 40))
int_year <- rep(int_year_i, each = length(unique(data$year)))

# Define data_stag
data_stag <- data.frame(int_year, data)

# Organizing the dataset with as.PanelMLCM
newdata <- as.PanelMLCM(y = data_stag[, "Y"], timevar = data_stag[, "year"], id = data_stag[, "ID"],
                        int_date = data_stag[, "int_year"],
                        x = data_stag[, !(names(data_stag) \%in\% c("Y", "ID", "year", "int_year"))], y.lag = 2)

# Panel Cross Validation in a staggered setting with different ML methods
pcv_multi <- PanelCrossValidationStag(data = newdata, ML_methods = list(enet, linreg, pls))

# ATE estimation
causal <- StagMLCM(data = newdata, int_date = "int_date", inf_type = "block", nboot = 10, PCV = pcv_multi)

}
