% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MLCM}
\alias{plot.MLCM}
\title{Plot of estimated causal effects}
\usage{
\method{plot}{MLCM}(x, type = c("ate", "cate"), ...)
}
\arguments{
\item{x}{Object returned from a previous call to \code{MLCM()}}

\item{type}{Character string indicating the plot to be produced. Possible values
in \code{c("ate", "cate")}}

\item{...}{Optional arguments from the \code{rpart.plot()} function to be used while
plotting the regression tree. See \code{help(rpart.plot)} for details.}
}
\value{
NULL, invisibly
}
\description{
Plot of estimated causal effects
}
\details{
The option \code{"ate"} plots the ATE at each point in time after the intervention.
Note that the effects plotted before the introduction of the policy are placebo effects (they
should be non-significant effects centered around zero otherwise either the intervention has
some anticipation effects or the model is misspecified). The option \code{"cate"} draws
the estimated regression tree using the \code{rpart.plot()} function from the \code{rpart.plot}
package. Users can use the \code{...} argument for full flexibility in drawing CATE regression
by exploiting all arguments of \code{rpart.plot()}.
}
