% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.StagMLCM}
\alias{plot.StagMLCM}
\title{Plot of estimated causal effects in staggered adoption setting}
\usage{
\method{plot}{StagMLCM}(x, type = c("global", "cohort", "time"))
}
\arguments{
\item{x}{Object returned from a previous call to \code{StagMLCM()}}

\item{type}{Character string indicating the plot to be produced. Possible values
in \code{c("global", "cohort", "time")}}
}
\value{
NULL, invisibly
}
\description{
Plot of estimated causal effects in staggered adoption setting
}
\details{
The option \code{"global"} plots the global ATE, i.e., a weighted average (by cohort size)
across all treatment groups of the temporal average causal effect of the policy. The option
\code{"cohort} plots the ATE over the entire post-intervention period separately for the
different cohorts. The option \code{"time} draws a weighted average (by group size) of the
across all treatment groups of the estimated ATEs in each time point after the intervention.
Note that the effects plotted before the introduction of the policy are placebo effects (they
should be non-significant effects centered around zero otherwise either the intervention has
some anticipation effects or the model is misspecified).
}
