% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcm.R
\name{StagMLCM}
\alias{StagMLCM}
\title{Machine Learning Control Method with staggered adoption}
\usage{
StagMLCM(
  data,
  int_date,
  inf_type = "block",
  y = NULL,
  timevar = NULL,
  id = NULL,
  y.lag = 0,
  nboot = 1000,
  pcv_block = 1,
  metric = "RMSE",
  default_par = list(gbm = list(depth = c(1, 2, 3), n.trees = c(500, 1000), shrinkage =
    seq(0.01, 0.1, by = 0.02), n.minobsinnode = c(10, 20)), rf = list(ntree = 500), pls =
    list(ncomp = c(1:5)), lasso = list(fraction = seq(0.1, 0.9, by = 0.1))),
  PCV = NULL,
  alpha = 0.05,
  fe = FALSE
)
}
\arguments{
\item{data}{A panel dataset in long form, having one column for the time variable, one column for the units'
unique IDs, one column for the outcome variable and one or more columns for the covariates. For staggered
adoption settings, the dataset must also have a column for the intervention dates.}

\item{int_date}{Name of the column containing the dates of the interventions.}

\item{inf_type}{Character, type of inference to be performed. Possible choices are 'classic', 'block', 'bc classic', 'bc block', 'bca'. Defaults to 'block'.}

\item{y}{Character, name of the column containing the outcome variable. It can be omitted for \code{PanelMLCM} objects.}

\item{timevar}{Character, name of the column containing the time variable. It can be omitted for \code{PanelMLCM} objects.}

\item{id}{Character, name of the column containing the ID's. It can be omitted for \code{PanelMLCM} objects.}

\item{y.lag}{Optional, number of lags of the dependent variable to include in the model. Defaults to zero.}

\item{nboot}{Number of bootstrap replications, defaults to 1000.}

\item{pcv_block}{Number of pre-intervention times to block for panel cross validation. Defaults to 1, see Details.}

\item{metric}{Character, the performance metric that should be used to select the optimal model.
Possible choices are either \code{"RMSE"} (the default) or \code{"Rsquared"}.}

\item{default_par}{List of parameters for the default ML algorithms. See the Details of \code{PanelCrossValidation}.}

\item{PCV}{Optional, list returned from a previous call to \code{PanelCrossValidation} or \code{PanelCrossValidationStag}.}

\item{alpha}{Confidence interval level to report for the ATE. Defaulting to 0.05 for a two sided
95\% confidence interval.}

\item{fe}{Logical, whether to include fixed effects dummy variables. Defaults to false.}
}
\value{
A list with the following components:
\itemize{
  \item \code{best_method}: the best-performing ML algorithm as selected by the PCV routine
  \item \code{int_date}: a vector with the intervention dates
  \item \code{global_ate}: a list with the estimated global ATE and its confidence interval
  \item \code{tempavg_individual}: a list with the estimated individual effects (averaged over
                                   the post-intervention periods) and their confidence intervals
  \item \code{group_ate}: a list with the estimated ATE for each treatment cohort and their
                          confidence intervals
  \item \code{groupavg}: a list with the estimated ATE for each period following the intervention and
                         their confidence intervals
}
}
\description{
This function adapts the MLCM approach to a staggered adoption setting. It takes as
input a panel dataset where groups of units are treated or shocked at
different times. It then performs the Panel Cross Validation (PCV) by comparing
the predictive performance of several Machine Learning (ML) methods in the pre-intervention
periods and then outputs the estimated effect of the policy. Three main causal estimands
are computed (see Details) and their confidence intervals are estimated by
bootstrap.
}
\details{
This function extends the MLCM method in a staggered adoption settings where groups of units are treated or shocked at
different times, e.g., one group of units becomes treated starting from 2022, another group of unit becomes treated
in 2024 and so on. In this case, the \code{data} must include a column with the units' intervention times.

The target causal estimand is the global ATE estimated as follows: a counterfactual is forecasted for
the units in each treatment group or cohort. If those groups are treated for more than one time period, a temporal average ATE
is calculated for each group and these are averaged together (a weighted average based on group size is used).
The result is the global effect of the policy or shock on all units that were treated.

In addition, the function computes an ATE for each cohort as well as a group-average effect of the policy at each point
in time following the intervention and the ATEs are averaged across the cohorts by their size (note that if the first cohort
becomed treated in 2022 and the others start being treated from 2023, the group-average ATE in 2022 is just the ATE of the
first cohort).
}
\examples{

### Example 1. Estimating ATE in staggered settings (with default ML methods)

# Assume the following intervention dates
int_year_i <- c(rep(2019, times = 60), rep(2020, times = 40))
int_year <- rep(int_year_i, each = length(unique(data$year)))

# Define data_stag
data_stag <- data.frame(int_year, data)

# Estimation
fit <- StagMLCM(data = data_stag, y = "Y", timevar = "year", id = "ID", int_date = "int_year",
            inf_type = "block", nboot = 10, y.lag = 2)

# Plotting
g <- plot(fit, type = "cohort")
g[[1]] + theme_bc()
g[[2]] + theme_bc()
}
