% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlcm.R
\name{as.PanelMLCM}
\alias{as.PanelMLCM}
\title{Structuring the panel dataset}
\usage{
as.PanelMLCM(y, x, timevar, id)
}
\arguments{
\item{y}{Character, name of the column containing the outcome variable.}

\item{x}{Matrix or data.frame of covariates to include in the model.}

\item{timevar}{Character, name of the column containing the time variable.}

\item{id}{Character, name of the column containing the ID's.}
}
\value{
An object of class 'data.frame' and 'PanelMLCM'.
}
\description{
This function takes as input the panel dataset given by the user and changes the
ordering and the names of the columns to obtain another data.frame of class 'PanelMLCM'
to be used by the function 'MLCM'. The function is for internal use (not exported).
}
