% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelcv.R
\name{PanelCrossValidationMulti}
\alias{PanelCrossValidationMulti}
\title{Panel Cross Validation Staggered}
\usage{
PanelCrossValidationMulti(
  data,
  pcv_block = 1,
  metric = "RMSE",
  trControl = NULL,
  ML_methods = NULL
)
}
\arguments{
\item{data}{A 'PanelMLCM' object from a previous call to \code{as.PanelMLCM}.}

\item{pcv_block}{Number of pre-intervention times to block for panel cross validation. Defaults to 1, see Details.}

\item{metric}{Character, the performance metric that should be used to select the optimal model.
Possible choices are either \code{"RMSE"} (the default) or \code{"Rsquared"}.}

\item{trControl}{Optional, used to customize the training step. It must be the output from a call to \code{trainControl} from the \code{caret} package.}

\item{ML_methods}{Optional list of ML methods to be used as alternatives to the default methods. Each method must be supplied
as a named list of two elements: a character defining the method name from all the ones available in \code{caret}
and the grid of parameter values to tune via the panel cross validation. See Details and the examples for additional explanations.#'}
}
\value{
For each intervention group, a list of class \code{train} with the best-performing ML method.
}
\description{
Panel Cross Validation in a staggered adoption setting, i.e., when there are groups of units
treated at different times.
}
\details{
The function works in the same way as \code{PanelCrossValidation} as it repeats independently the PCV procedure for each group of unit defined by
their treatment date. Note that the \code{int_date} argument must be a column of \code{data}.
}
\examples{

### Example 1. Using the default ML methods in a staggered setting

# Assume the following intervention dates
int_year_i <- c(rep(2019, times = 60), rep(2020, times = 40))
int_year <- rep(int_year_i, each = length(unique(data$year)))

# Define data2
data2 <- data.frame(int_year, data)

# Organizing the dataset with as.PanelMLCM
newdata <- as.PanelMLCM(y = data2[, "Y"], timevar = data2[, "year"], id = data[, "ID"],
                        int_date = data2[, "int_year"],
                        x = data2[, !(names(data) \%in\% c("Y", "ID", "year", "int_year"))], y.lag = 1)

# Panel Cross Validation in a staggered setting
pcv <- PanelCrossValidationMulti(data = newdata)

}
